requirements.txt
streamlit
requests
pandas
plotly
pillow